Poprawienie struktury ekranow
