+++ DanteBlackjack/Screen.py
@@ -1,6 +1,7 @@
 import pygame
 pygame.init()
 from Button import *
+from InitBox import *
 
 class Screen:
     def Start(self, window, choice):
