+++ DanteBlackjack/main.py
@@ -1,46 +1,46 @@
-import pygame
-pygame.init()
-from EkranStartowyDante import *
-from EkranStartuGry import *
-from Lobby_Screen import *
-from EkranLogowania import *
+from DanteStartScreen import *
+from DanteBlackjackStartScreen import *
+from LobbyScreen import *
+from LoginDanteScreen import *
+
 
 window = pygame.display.set_mode(resolutions[choice])
 
-ekran_logowania = EkranLogowania()
-ekran_startowy_dante = EkranStartowyDante()
-ekran_startu_gry = EkranStartuGry()
-ekran_wyboru_stolu = Lobby_Screen()
-current_screen = ekran_logowania
+login_dante_screen = LoginDanteScreen()
+dante_start_screen = DanteStartScreen()
+dante_blackjack_start_screen = DanteBlackjackStartScreen()
+lobby_screen = LobbyScreen()
+current_screen = login_dante_screen
 
 running = True
 while running:
+    current_screen.Start(window, choice)
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
             running = False
-        if(current_screen == ekran_logowania):
+        elif event.type == pygame.MOUSEBUTTONDOWN and event.button == 1:
+            #ekran logowania
+            if button_log.tool_click_left() and current_screen == login_dante_screen:
+                username.status_set_0()
+                current_screen = dante_start_screen
+            if username.tool_click_left() and current_screen == login_dante_screen:
+                username.status_set_1()
+                password.status_set_0()
+            if password.tool_click_left() and current_screen == login_dante_screen:
+                password.status_set_1()
+                username.status_set_0()
+            #ekran startowy dante
+            if button_0.tool_click_left() and current_screen == dante_start_screen:
+                current_screen = dante_blackjack_start_screen
+            #ekran startu gry
+            if button_1.tool_click_left() and current_screen == dante_blackjack_start_screen:
+                current_screen = dante_start_screen
+            if button_2.tool_click_left() and current_screen == dante_blackjack_start_screen:
+                current_screen = lobby_screen
+            #ekran wybor stolu
+            if lobby_to_Menu_button.tool_click_left() and current_screen == lobby_screen:
+                current_screen = dante_blackjack_start_screen
+        #ekran logowania
+        if(current_screen == login_dante_screen):
             username.writing(event)
             password.writing(event)
-    current_screen.Start(window, choice)
-    #ekran logowania
-    if button_log.tool_click_left() and current_screen == ekran_logowania:
-        username.status_set_0()
-        current_screen = ekran_startowy_dante
-    if username.tool_click_left() and current_screen == ekran_logowania:
-        username.status_set_1()
-        password.status_set_0()
-    if password.tool_click_left() and current_screen == ekran_logowania:
-        password.status_set_1()
-        username.status_set_0()
-    #ekran startowy dante
-    if button_0.tool_click_left() and current_screen == ekran_startowy_dante:
-        current_screen = ekran_startu_gry
-    #ekran startu gry
-    if button_1.tool_click_left() and current_screen == ekran_startu_gry:
-        current_screen = ekran_startowy_dante
-    if button_2.tool_click_left() and current_screen == ekran_startu_gry:
-        current_screen = ekran_wyboru_stolu
-    #ekran wybor stolu
-    if lobby_to_Menu_button.tool_click_left() and current_screen == ekran_wyboru_stolu:
-        current_screen = ekran_startu_gry
-
